GWT-689: remove reference to java.net.URL: cannot be used client-side. Used blunt methods to separate base from query url
